requirements.txt
Django==5.0.2
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
markdown==3.5.2
python-frontmatter==1.1.0 